OutputFramerate

+++ src/args/argsParser.go
@@ -75,10 +75,10 @@ func ParseArgs(args []string) (config multiplexer.MultiplexerConfig, err error) 
 	outputHeightPtr := CommandLine.Int("output_height", -1, OutputHeigthUsage)
 	modePtr := CommandLine.String("mode", "", ModeUsage) // grid OR motion
 	modeGridPtr := CommandLine.String("grid_dimension", "", GridUsage)
-	credentialsPtr := CommandLine.String("use_auth", "", CredentialsUsage)      // grid OR motion
-	logTimePtr := CommandLine.Bool("log_time", false, LogTimeUsage)             // optional flag
-	inputDelayPtr := CommandLine.Int("min_input_delay", 0, MinInputDelay)       // optional flag
-	outputFrameRate := CommandLine.Int("output_framerate", -1, OutputFramerate) // optional flag
+	credentialsPtr := CommandLine.String("use_auth", "", CredentialsUsage)          // grid OR motion
+	logTimePtr := CommandLine.Bool("log_time", false, LogTimeUsage)                 // optional flag
+	inputDelayPtr := CommandLine.Int("min_input_delay", 0, MinInputDelay)           // optional flag
+	outputFrameRate := CommandLine.Float64("output_framerate", -1, OutputFramerate) // optional flag
 
 	//---parse the command line into the defined flags---
 	CommandLine.Parse(args[1:])
